Frontend: Avoid deadlock in Nix>>build:
Work around a bug in the Pharo subprocess support to avoid deadlocking
on a Nix build that produces a lot of output.

The root problem is that Pharo is only reading from stdout, and not
stderr, while waiting for a subprocess to terminate. Therefore if the
subprocess fills up its stderr then it will become deadlocked because
Pharo is responsible for draining the pipe but did not do so.

The deadlocked process looks like this in strace:

    write(2, ...

i.e. a write to its stderr file (fd 2) is waiting for space to become available.

This is especially problematic for running Nix builds because the
deadlocked process will also hold locks on the Nix store which can
cause cascading deadlocks to other processes.